test(cli): cover TINA_EXTERNAL_PKGS parsing and finalize config bundling updates
Agent-Logs-Url: https://github.com/link89/tinacms/sessions/1352c2cf-c864-4e9f-b28d-5ea8c407cb2d
